# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gerenciamento de Pedidos AAAAAAAA",
    "description": "Desafio Tech Challange FIAP",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/pedidos": {
      "get": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Busca pedidos por status",
        "operationId": "buscarPedidos",
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseDeDadosException"
                }
              }
            }
          },
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PedidoResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Cadastrar um pedido",
        "operationId": "post_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastrarPedidoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseDeDadosException"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PedidoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pagamentos/finalizar": {
      "post": {
        "tags": [
          "pagamento-controller"
        ],
        "summary": "Finaliza o pagamento com status APROVADO ou REPROVADO",
        "operationId": "finalizarPagamento",
        "parameters": [
          {
            "name": "data.id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagamentoCriadoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseDeDadosException"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/pedidos/status": {
      "patch": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Alterar status de um pedido",
        "operationId": "alterarStatusPedido",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlterarStatusPedidoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseDeDadosException"
                }
              }
            }
          }
        }
      }
    },
...
    "/pedidos/{id}/pagamento/status": {
      "get": {
        "tags": [
          "pedido-controller"
        ],
        "summary": "Consultar status do pagamento de um pedido",
        "operationId": "consultarStatusPagamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseDeDadosException"
                }
              }
            }
          },
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagamentoStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseDeDadosException": {
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "cause": {
            "type": "object",
            "properties": {
              "stack_trace": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "class_loader_name": {
                      "type": "string"
                    },
                    "module_name": {
                      "type": "string"
                    },
                    "module_version": {
                      "type": "string"
                    },
                    "method_name": {
                      "type": "string"
                    },
                    "file_name": {
                      "type": "string"
                    },
                    "line_number": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "native_method": {
                      "type": "boolean"
                    },
                    "class_name": {
                      "type": "string"
                    }
                  }
                }
              },
              "message": {
                "type": "string"
              },
              "suppressed": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "stack_trace": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "class_loader_name": {
                            "type": "string"
                          },
                          "module_name": {
                            "type": "string"
                          },
                          "module_version": {
                            "type": "string"
                          },
                          "method_name": {
                            "type": "string"
                          },
                          "file_name": {
                            "type": "string"
                          },
                          "line_number": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "native_method": {
                            "type": "boolean"
                          },
                          "class_name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "localized_message": {
                      "type": "string"
                    }
                  }
                }
              },
              "localized_message": {
                "type": "string"
              }
            }
          },
          "stack_trace": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "class_loader_name": {
                  "type": "string"
                },
                "module_name": {
                  "type": "string"
                },
                "module_version": {
                  "type": "string"
                },
                "method_name": {
                  "type": "string"
                },
                "file_name": {
                  "type": "string"
                },
                "line_number": {
                  "type": "integer",
                  "format": "int32"
                },
                "native_method": {
                  "type": "boolean"
                },
                "class_name": {
                  "type": "string"
                }
              }
            }
          },
          "suppressed": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "stack_trace": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "class_loader_name": {
                        "type": "string"
                      },
                      "module_name": {
                        "type": "string"
                      },
                      "module_version": {
                        "type": "string"
                      },
                      "method_name": {
                        "type": "string"
                      },
                      "file_name": {
                        "type": "string"
                      },
                      "line_number": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "native_method": {
                        "type": "boolean"
                      },
                      "class_name": {
                        "type": "string"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                },
                "localized_message": {
                  "type": "string"
                }
              }
            }
          },
          "localized_message": {
            "type": "string"
          }
        }
      },
      "Imagem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "caminho": {
            "type": "string"
          }
        }
      },
...
      "Produto": {
        "required": [
          "disponivel",
          "excluido",
          "tempo_preparo",
          "valor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "categoria": {
            "type": "string",
            "enum": [
              "LANCHE",
              "ACOMPANHAMENTO",
              "BEBIDA",
              "SOBREMESA"
            ]
          },
          "valor": {
            "type": "number"
          },
          "tempo_preparo": {
            "type": "integer",
            "format": "int64"
          },
          "disponivel": {
            "type": "boolean"
          },
          "excluido": {
            "type": "boolean"
          },
          "imagem": {
            "$ref": "#/components/schemas/Imagem"
          }
        }
      },
...
      "CadastrarPedidoProdutoRequest": {
        "required": [
          "produto_id",
          "quantidade"
        ],
        "type": "object",
        "properties": {
          "produto_id": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "integer",
            "format": "int64"
          },
          "comentario": {
            "type": "string"
          }
        }
      },
      "CadastrarPedidoRequest": {
        "required": [
          "produto_ids"
        ],
        "type": "object",
        "properties": {
          "cliente_id": {
            "type": "integer",
            "format": "int64"
          },
          "produtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CadastrarPedidoProdutoRequest"
            }
          },
          "produto_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "Cliente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cpf": {
            "$ref": "#/components/schemas/Cpf"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "$ref": "#/components/schemas/Email"
          }
        }
      },
      "ClienteResponse": {
        "type": "object",
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cpf": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "Cpf": {
        "required": [
          "numero"
        ],
        "type": "object",
        "properties": {
          "numero": {
            "type": "string"
          }
        }
      },
      "Email": {
        "required": [
          "endereco"
        ],
        "type": "object",
        "properties": {
          "endereco": {
            "type": "string"
          }
        }
      },
      "Item": {
        "required": [
          "quantidade"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "integer",
            "format": "int64"
          },
          "comentario": {
            "type": "string"
          },
          "produto": {
            "$ref": "#/components/schemas/Produto"
          },
          "valor_pago": {
            "type": "number"
          }
        }
      },
      "Pagamento": {
        "required": [
          "data_hora",
          "qr_code",
          "status",
          "valor_total"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "APROVADO",
              "REPROVADO"
            ]
          },
          "qr_code": {
            "type": "string"
          },
          "valor_total": {
            "type": "number"
          }
        }
      },
      "PagamentoResponse": {
        "required": [
          "data_hora",
          "qr_code",
          "status",
          "valor_total"
        ],
        "type": "object",
        "properties": {
          "pagamento": {
            "$ref": "#/components/schemas/Pagamento"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "APROVADO",
              "REPROVADO"
            ]
          },
          "qr_code": {
            "type": "string"
          },
          "valor_total": {
            "type": "number"
          }
        }
      },
      "Pedido": {
        "required": [
          "data_hora",
          "items",
          "status"
        ],
        "type": "object",
        "properties": {
          "numero": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "RECEBIDO",
              "EM_PREPARACAO",
              "PRONTO",
              "FINALIZADO"
            ]
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "pagamento": {
            "$ref": "#/components/schemas/Pagamento"
          },
          "tempo_espera_minutos": {
            "type": "integer",
            "format": "int64"
          },
          "valor_total": {
            "type": "number"
          }
        }
      },
      "PedidoProdutoResponse": {
        "required": [
          "quantidade"
        ],
        "type": "object",
        "properties": {
          "pedidoProduto": {
            "$ref": "#/components/schemas/Item"
          },
          "nome": {
            "type": "string"
          },
          "valor_pago": {
            "type": "number"
          },
          "quantidade": {
            "type": "integer",
            "format": "int64"
          },
          "comentario": {
            "type": "string"
          }
        }
      },
      "PedidoResponse": {
        "required": [
          "data_hora",
          "produtos",
          "status"
        ],
        "type": "object",
        "properties": {
          "pedido": {
            "$ref": "#/components/schemas/Pedido"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "RECEBIDO",
              "EM_PREPARACAO",
              "PRONTO",
              "FINALIZADO"
            ]
          },
          "tempo_espera_minutos": {
            "type": "integer",
            "format": "int64"
          },
          "numero": {
            "type": "string"
          },
          "cliente": {
            "$ref": "#/components/schemas/ClienteResponse"
          },
          "produtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PedidoProdutoResponse"
            }
          },
          "pagamento": {
            "$ref": "#/components/schemas/PagamentoResponse"
          }
        }
      },
      "DataRequest": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "PagamentoCriadoRequest": {
        "required": [
          "action",
          "api_version",
          "data",
          "date_created",
          "id",
          "live_mode",
          "type",
          "user_id"
        ],
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "api_version": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DataRequest"
          },
          "date_created": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "live_mode": {
            "type": "boolean"
          },
          "type": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        }
      },
...
      "AlterarStatusPedidoRequest": {
        "required": [
          "pedido_id",
          "status"
        ],
        "type": "object",
        "properties": {
          "pedido_id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "RECEBIDO",
              "EM_PREPARACAO",
              "PRONTO",
              "FINALIZADO"
            ]
          }
        }
      },
      "PagamentoStatusResponse": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "APROVADO",
              "REPROVADO"
            ]
          }
        }
      }
    }
  }
}